Enviando arquivos de mais um modulo

+++ app/Http/Controllers/Admin/PlanController.php
@@ -63,11 +63,16 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with('details')->where('url', $url)->first();
 
         if (!$plan)
             return redirect()->back();
 
+        //dd($plan);//->details->count();
+
+        if($plan->details->count() > 0){
+            return redirect()->back()->with('error','Exitem detalhes vinculados ao plano, impossível deletar registro.');
+        }
         $plan->delete();
 
         return redirect()->route('plans.index');
